@@ -2,44 +2,39 @@
   "name": "traintrack-webclient",
   "dependencies": {
     "@types/history": "^4.7.11",
-    "@types/node": "^12.20.55",
-    "@types/react": "^16.14.40",
-    "@types/react-dom": "^16.9.18",
-    "@types/react-helmet": "^5.0.21",
+    "@types/node": "^18.15.12",
+    "@types/react": "^18.0.37",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-helmet": "^6.1.6",
     "@types/react-redux": "^7.1.25",
-    "@types/react-router-dom": "^4.3.5",
+    "@types/react-router-dom": "^6.10.0",
     "es6-map": "^0.1.5",
     "google-protobuf": "^3.21.2",
     "immutable": "^4.3.0",
     "moment": "^2.29.4",
-    "protobufjs": "^6.11.3",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-g-analytics": "^0.4.2",
-    "react-helmet": "^5.2.1",
-    "react-redux": "^7.2.9",
+    "protobufjs": "^7.2.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-redux": "^8.0.5",
     "react-redux-typescript": "^3.0.0-rc.3",
-    "react-router-dom": "^5.3.4",
     "react-router-redux": "^4.0.8",
     "redux": "^4.2.1",
     "redux-thunk": "^2.4.2"
   },
   "devDependencies": {
-    "awesome-typescript-loader": "^5.2.1",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-runtime": "^6.26.0",
-    "compression-webpack-plugin": "^1.1.12",
-    "query-string": "^6.14.1",
-    "source-map-loader": "^0.2.4",
-    "typescript": "^3.9.10",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12"
+    "@babel/core": "^7.21.4",
+    "@babel/plugin-transform-runtime": "^7.21.4",
+    "@babel/preset-env": "^7.21.4",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/runtime": "^7.21.0",
+    "compression-webpack-plugin": "^10.0.0",
+    "query-string": "^8.1.0",
+    "react-router-dom": "^6.10.0",
+    "source-map-loader": "^4.0.1",
+    "ts-loader": "^9.4.2",
+    "typescript": "^5.0.4",
+    "webpack": "^5.80.0",
+    "webpack-cli": "^5.0.1"
   }
 }
