@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'babel-polyfill';
+//import 'babel-polyfill';
 
 import * as moment from "moment";
 
